refactor(teams): split headless.rs into focused config, run, and resume submodules [OPE-440]
Split the 485-line headless.rs into a module directory with clear boundaries:
- config.rs: team loading and headless context creation helpers
- run.rs: run_headless, run_headless_with_model, run_headless_with_project
- resume.rs: resume_headless, find_parent_work_item
- tests.rs: all unit tests (preserved as-is)
- mod.rs: module declarations and public re-exports

All 9 headless tests pass. No public API changes.
