+++ crates/opengoose-teams/src/headless/mod.rs
@@ -1,0 +1,10 @@
+mod config;
+pub(crate) mod resume;
+pub(crate) mod run;
+
+#[cfg(test)]
+mod tests;
+
+pub use config::HeadlessConfig;
+pub use resume::resume_headless;
+pub use run::run_headless;
